requirements.txt
# Requirements for SharePoint Excel file access with modern authentication

# MSAL for Azure AD authentication
msal==1.24.1

# Office365 REST client library
Office365-REST-Python-Client==2.5.3

# HTTP requests library
requests==2.31.0

# Data manipulation and Excel file handling
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Additional utilities
python-dotenv==1.0.0  # For environment variable management
urllib3==2.0.7
certifi==2023.7.22